+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -5,8 +5,7 @@ CHANGELOG
 -----
 
  * added `%env(trim:...)%` processor to trim a string value
- * added `%env(default:...)%` processor to fallback to a default value
- * added `%env(nullable:...)%` processor to allow empty variables to be processed as null values
+ * added `%env(default:param_name:...)%` processor to fallback to a parameter or to null when using `%env(default::...)%`
  * added support for deprecating aliases
  * made `ContainerParametersResource` final and not implement `Serializable` anymore
  * added ability to define an index for a tagged collection